refactor: `hookCollector` add `hookCalls` collect, remove `cost` field
