+++ packages/core/src/hook/index.ts
@@ -1,5 +1,5 @@
 export * from "./hierarchy";
+export * from "./hook";
 export * from "./hook-call";
 export * from "./hook-collector";
-export * from "./hook-kind";
 export * from "./hook-name";
